Rename so that flow_blackoil_tpfa becomes flow_blackoil.
